Issue #181 Security fix and roles check in getting user groups fix

+++ LearnWithMentor/Controllers/GroupController.cs
@@ -9,6 +9,8 @@ using LearnWithMentorBLL.Services;
 using System.Web.Http.Tracing;
 using LearnWithMentor.Log;
 using System.Data.Entity.Core;
+using System.Web;
+using System.Security.Claims;
 
 namespace LearnWithMentor.Controllers
 {
@@ -44,9 +46,13 @@ namespace LearnWithMentor.Controllers
         {
             var allGroups = groupService.GetGroupsByMentor(id);
             if (allGroups != null)
+            {
                 return Request.CreateResponse(HttpStatusCode.OK, allGroups);
+            }
             else
+            {
                 return Request.CreateErrorResponse(HttpStatusCode.NotFound, $"No groups for the mentor in database. (mentorId = {id})");
+            }
         }
 
         /// <summary>
@@ -60,9 +66,13 @@ namespace LearnWithMentor.Controllers
         {
             var group = groupService.GetGroupById(id);
             if (group != null)
+            {
                 return Request.CreateResponse(HttpStatusCode.OK, group);
+            }
             else
+            {
                 return Request.CreateErrorResponse(HttpStatusCode.NotFound, $"There isn't group with id = {id}");
+            }
         }
 
         /// <summary>
@@ -78,9 +88,13 @@ namespace LearnWithMentor.Controllers
             {
                 var group = groupService.GetPlans(id);
                 if (group != null)
+                {
                     return Request.CreateResponse(HttpStatusCode.OK, group);
+                }
                 else
+                {
                     return Request.CreateErrorResponse(HttpStatusCode.NoContent, $"There no plans in this group.");
+                }
             }
             catch (EntityException e)
             {
@@ -102,9 +116,13 @@ namespace LearnWithMentor.Controllers
             {
                 var group = groupService.GetUsers(id);
                 if (group != null)
+                {
                     return Request.CreateResponse(HttpStatusCode.OK, group);
+                }
                 else
+                {
                     return Request.CreateErrorResponse(HttpStatusCode.NoContent, $"There are no users in the group.");
+                }
             }
             catch (EntityException e)
             {
@@ -118,20 +136,26 @@ namespace LearnWithMentor.Controllers
         /// </summary>
         /// <param name="groupId"></param>
         /// <returns></returns>
+        [Authorize(Roles = "Mentor")]
         [HttpGet]
         [Route("api/group/{groupId}/users/notingroup")]
         public HttpResponseMessage GetUsersNotInCurrentGroup(int groupId)
         {
             var group = groupService.GetUsersNotInGroup(groupId);
             if (group != null)
+            {
                 return Request.CreateResponse(HttpStatusCode.OK, group);
+            }
             else
+            {
                 return Request.CreateErrorResponse(HttpStatusCode.NotFound, $"There isn't users outside of the group id = {groupId}");
+            }
         }
 
         /// <summary>
         /// Returns all plans not used in current group.
         /// </summary>
+        [Authorize(Roles = "Mentor")]
         [HttpGet]
         [Route("api/plan/notingroup/{groupId}")]
         public HttpResponseMessage GetPlansNotUsedInCurrentGroup(int groupId)
@@ -159,7 +183,9 @@ namespace LearnWithMentor.Controllers
             try
             {
                 if (!ModelState.IsValid)
+                {
                     return Request.CreateErrorResponse(HttpStatusCode.BadRequest, ModelState);
+                }
                 bool success = groupService.AddGroup(group);
                 if (success)
                 {
@@ -190,6 +216,13 @@ namespace LearnWithMentor.Controllers
         {
             try
             {
+                var identity = HttpContext.Current.User.Identity as ClaimsIdentity;
+                var currentUserId = int.Parse(identity.FindFirst("Id").Value);
+                var mentorId = groupService.GetMentorIdByGroup(id);
+                if (mentorId != currentUserId)
+                {
+                    return Request.CreateErrorResponse(HttpStatusCode.Unauthorized, "Authorization denied.");
+                }
                 bool success = groupService.AddUsersToGroup(userId, id);
                 if (success)
                 {
@@ -220,6 +253,13 @@ namespace LearnWithMentor.Controllers
         {
             try
             {
+                var identity = HttpContext.Current.User.Identity as ClaimsIdentity;
+                var userId = int.Parse(identity.FindFirst("Id").Value);
+                var mentorId = groupService.GetMentorIdByGroup(id);
+                if (mentorId != userId)
+                {
+                    return Request.CreateErrorResponse(HttpStatusCode.Unauthorized, "Authorization denied.");
+                }
                 bool success = groupService.AddPlansToGroup(planId, id);
                 if (success)
                 {
@@ -264,7 +304,9 @@ namespace LearnWithMentor.Controllers
                 string[] lines = searchKey.Split(new char[] { ' ' }, StringSplitOptions.RemoveEmptyEntries);
                 var plansList = groupService.SearchPlansNotUsedInGroup(lines, groupId);
                 if (plansList == null)
+                {
                     return Request.CreateErrorResponse(HttpStatusCode.NoContent, "This plan does not exist.");
+                }
                 return Request.CreateResponse(HttpStatusCode.OK, plansList);
             }
             catch (EntityException e)
@@ -292,7 +334,9 @@ namespace LearnWithMentor.Controllers
                 string[] lines = searchKey.Split(new char[] { ' ' }, StringSplitOptions.RemoveEmptyEntries);
                 var usersList = groupService.SearchUserNotInGroup(lines, groupId);
                 if (usersList == null)
+                {
                     return Request.CreateErrorResponse(HttpStatusCode.NoContent, "This user does not exist.");
+                }
                 return Request.CreateResponse(HttpStatusCode.OK, usersList);
             }
             catch (EntityException e)
@@ -314,6 +358,13 @@ namespace LearnWithMentor.Controllers
         {
             try
             {
+                var identity = HttpContext.Current.User.Identity as ClaimsIdentity;
+                var id = int.Parse(identity.FindFirst("Id").Value);
+                var mentorId = groupService.GetMentorIdByGroup(groupId);
+                if (mentorId != id)
+                {
+                    return Request.CreateErrorResponse(HttpStatusCode.Unauthorized, "Authorization denied.");
+                }
                 bool successfullyRemoved = groupService.RemoveUserFromGroup(groupId, userToRemoveId);
                 if (successfullyRemoved)
                 {
@@ -367,18 +418,26 @@ namespace LearnWithMentor.Controllers
         /// <param name="userId"></param>
         /// <returns></returns>
         [HttpGet]
-        [Route("api/group/user/{userId}/groups")]
-        public HttpResponseMessage GetUserGroups(int userId)
+        [Route("api/group/mygroups")]
+        public HttpResponseMessage GetUserGroups()
         {
             try
             {
+                var identity = HttpContext.Current.User.Identity as ClaimsIdentity;
+                var userId = int.Parse(identity.FindFirst("Id").Value);
                 if (!userService.ContainsId(userId))
+                {
                     return Request.CreateErrorResponse(HttpStatusCode.NoContent, $"There are no users with id = {userId}");
+                }
                 if (groupService.GroupsCount() == 0)
+                {
                     return Request.CreateErrorResponse(HttpStatusCode.NoContent, $"There are no groups in database.");
+                }
                 var groups = groupService.GetUserGroups(userId);
                 if (groups == null)
+                {
                     return Request.CreateErrorResponse(HttpStatusCode.BadRequest, $"There are no groups for this user");
+                }
                 return Request.CreateResponse(HttpStatusCode.OK, groups);
             }
             catch (EntityException e)
